refactor executors module
